+++ lib/screens/main_page.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+import '../widgets/bottom_navigation.dart';
+import 'search_screen.dart';
+import 'setting.dart';
+
+class MainPage extends StatefulWidget {
+  @override
+  _MainPageState createState() => _MainPageState();
+}
+
+class _MainPageState extends State<MainPage> {
+  int _selectedIndex = 0;
+
+  static const List<Widget> _widgetOptions = <Widget>[
+    SearchPage(),
+    SettingPage(),
+  ];
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: _widgetOptions.elementAt(_selectedIndex),
+      ),
+      bottomNavigationBar: BottomNavigation(
+        currentIndex: _selectedIndex,
+        onItemSelected: _onItemTapped,
+      ),
+    );
+  }
+}
